pyproject.toml
[project]
name = "hello-slither"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "slither-analyzer>=0.11.3",
]
